% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_bootstrap_ci.R
\name{growth_boostrap_ci}
\alias{growth_boostrap_ci}
\title{Create bootstrap estimates and 95\% confidence intervals for mixed-effects
and/or least-squares models for each time-point}
\usage{
growth_boostrap_ci(
  data_frame,
  growth_model_object,
  growth_model_summary_list,
  boot_n_sim = 200,
  mix_boot_method = "case"
)
}
\arguments{
\item{data_frame}{A data frame object that at minimum contains three
variables:
\itemize{
\item cluster - a character type variable used to specify how observations
are nested or grouped by a particular cluster. Note if using a
least-squares model, please fill in all values of cluster with a single
dummy character string, do NOT leave blank.
\item time - a numeric type variable used for measuring time such as
minutes, hours, or days
\item growth_metric - a numeric type variable used for measuring growth
over time such as cell count or confluency
}}

\item{growth_model_object}{A saemix or nls type model object that is
created using \code{\link{growth_curve_model_fit}} when
return_summary = FALSE}

\item{growth_model_summary_list}{A list object created by the
\code{\link{growth_curve_model_fit}} function when return_summary = TRUE.}

\item{boot_n_sim}{A numeric value specifying the number of bootstrap
simulations to be performed. See \code{\link[saemix]{saemix.bootstrap}}
for mixed-effects models and \code{\link[nlraa]{predict_nls}} for
least-squares models. Defaults to 200.}

\item{mix_boot_method}{For mixed-effects models, a character string
specifying the bootstrap algorithm to use. Options include "case",
"residual", "parametric" or "conditional". Defaults to "case". See
\code{\link[saemix]{saemix.bootstrap}} for more details.}
}
\value{
An appended version of growth_model_summary_list with a fourth
data frame titled "boot_sim"
}
\description{
This function leverages the \code{\link[saemix]{saemix.bootstrap}} function
for mixed-effects models and the \code{\link[nlraa]{predict_nls}} function
for least-squares models to compute bootstrapped 95\% confidence intervals for
each time-point for graphical purposes. Estimates of the fixed-effects
values are calculated based on the median (50th percentile) of the
simulated bootstrap data, with the 95\% confidence interval constructed
from the 2.5th and 97.5th percentiles.
}
\examples{
# Fit an mixed-effects growth model to the data, return summary
# and bootstrap estimates (boot_n_sim set to 6 for speed)
exp_mixed_model_summary <- growth_curve_model_fit(
  data_frame = exp_mixed_data,
  function_type = "exponential",
  bootstrap_time = TRUE,
  boot_n_sim = 6
)
}
\seealso{
\code{\link{growth_curve_model_fit}}
}
