% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curve_model_fit.R
\name{growth_curve_model_fit}
\alias{growth_curve_model_fit}
\title{Fit a growth function using mixed-effects regression modeling}
\usage{
growth_curve_model_fit(
  data_frame,
  function_type = "exponential",
  model_type = "mixed",
  fixed_rate = TRUE
)
}
\arguments{
\item{data_frame}{A data frame object that at minimum contains three variables:
\itemize{
\item cluster - a character type variable used to specify how observations are nested or grouped by a particular cluster. Note if using a least-squares model, please fill in all values of cluster with a single dummy character string, do NOT leave blank.
\item time - a numeric type variable used for measuring time such as minutes, hours, or days
\item growth_metric - a numeric type variable used for measuring growth over time such as cell count or confluency
}}

\item{function_type}{A character string specifying the function for modeling the shape of the growth. Options include "exponential", "linear", "logistic", or "gompertz".}

\item{model_type}{A character string specifying the type of regression model to be used. If "mixed" a mixed-effects regression model will be used with fixed and random effects to account for clustering. If "least-squares". Defaults to "mixed".}

\item{fixed_rate}{A logical value specifying whether the rate constant of the function should be treated as a fixed effect (TRUE) or random effect (FALSE). Defaults to TRUE}
}
\value{
Returns a model object of class 'nlme' when a mixed-effects model is specified or a model object of class 'nls' if a least-squares model is specified.
}
\description{
'growth_curve_model_fit()' fits a mixed-effects model to a data frame based on a user-defined function to account for clustering.
}
\examples{
# Load example data (exponential data)
data(exp_mixed_data)
# Fit an mixed-effects growth model to the data
exp_mixed_model <- growth_curve_model_fit(data_frame = exp_mixed_data, function_type = "exponential")
# Summarize the data by creating a summary list object
exp_mixed_model_summary <- summarize_growth_model(data_frame = exp_mixed_data,
growth_model_object = exp_mixed_model,
model_type = "mixed",
function_type = "exponential")
# Create flextable object from the summary list object for documentation
exp_model_table <- growth_model_summary_table(growth_model_summary_list = exp_mixed_model_summary)
print(exp_model_table)
# Create growth vs time plot of data with fitted values (plot_type = 2), adjust aesthetics and parameters as desired
exp_growth_plot <- growth_vs_time_plot(growth_model_summary_list = exp_mixed_model_summary,
model_type = "mixed",
plot_type = 2)
print(exp_growth_plot)
# Check residuals and model assumptions
residual_diag_plot <- growth_model_residual_plots(growth_model_summary_list = exp_mixed_model_summary,
model_type = "mixed",
residual_type = "conditional")
print(residual_diag_plot)
# Create a plot of fixed-effects bootstrapped 95\%CI with the doubling time estimates from the exp_mixed_model annotated on the plot
exp_ci_plot <- growth_model_boot_ci_curve(growth_model_object = exp_mixed_model,
growth_model_summary_list = exp_mixed_model_summary,
model_type = "mixed",
annotate_value = "double_time",
return = "plot")
print(exp_ci_plot)
}
