% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_model_boot_ci_curve.R
\name{growth_model_boot_ci_curve}
\alias{growth_model_boot_ci_curve}
\title{Create bootstrapped confidence interval plot from fixed-effects of growth model}
\usage{
growth_model_boot_ci_curve(
  growth_model_object,
  growth_model_summary_list,
  model_type = "mixed",
  annotate_value = "double_time",
  return = "plot",
  growth_metric_name = "growth_metric",
  time_name = "time",
  plot_title = "Regression Model with 95\% Confidence Intervals",
  boot_n_sim = 1000,
  x_limits = c(NA, NA),
  n_x_axis_breaks = NULL,
  y_limits = c(NA, NA),
  n_y_axis_breaks = NULL,
  x_axis_text_size = 8,
  y_axis_text_size = 12,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  plot_title_size = 20,
  annotate_value_text_size = 6,
  ...
)
}
\arguments{
\item{growth_model_object}{The model object that is created using the growth_curve_model_fit()}

\item{growth_model_summary_list}{A list object created by the \code{\link{summarize_growth_model}} function.}

\item{model_type}{A character string specifying the model_type that was fit using the \code{\link{growth_curve_model_fit}} function. Options include either "mixed" or "least-squares. Defaults to "mixed".}

\item{annotate_value}{A character string specifying whether the doubling time estimate with 95\%CI or rate constant estimate with 95\%CI derived from the growth_model_object should be displayed on the plot.
Options include "NA" for no annotation, "double_time" for the doubling time with 95\%CI, or "rate_constant" for the rate constant with 95\%CI.
Defaults to "double_time".
Please note that when the growth model is a mixed-effects model, the 95\%CI annotation accounts for fixed and random effects, where as the 95\%CI plotted on the graph is generated from bootstrapping the fixed effects of the model only.}

\item{return}{A character string, options include "plot" for returning the plot generated from the function or "estimates" to return a data frame with the predicted bootstrap confidence values for each observation. Defaults to "plot".}

\item{growth_metric_name}{A character string for specifying the name of the growth metric (y-axis title) to be displayed on the plot. Defaults to "growth_metric".}

\item{time_name}{A character string for specifying the name of the time variable (x-axis title) to be displayed on the plot. Defaults to "time".}

\item{plot_title}{A character string for specifying the title to be displayed over the plot. Defaults to "Regression Model with 95\% Confidence Intervals".}

\item{boot_n_sim}{A numeric value specifying the number of simulations for \code{\link[nlraa]{predict_nlme}} or \code{\link[nlraa]{predict_nls}} to perform to create the confidence intervals. Defaults to 1000.}

\item{x_limits}{A numeric vector of length two providing limits for the x-axis. Use NA to refer to the existing minimum or maximum. Defaults to c(NA, NA). See \code{\link[ggplot2]{scale_x_continuous}}.}

\item{n_x_axis_breaks}{An integer specifying the number of major breaks for the x-axis. Defaults to NULL. See \code{\link[ggplot2]{scale_x_continuous}}.}

\item{y_limits}{A numeric vector of length two providing limits for the y-axis. Use NA to refer to the existing minimum or maximum. Defaults to c(NA, NA). See \code{\link[ggplot2]{scale_y_continuous}}.}

\item{n_y_axis_breaks}{An integer specifying the number of major breaks for the x-axis. Defaults to NULL. See \code{\link[ggplot2]{scale_y_continuous}}.}

\item{x_axis_text_size}{A numeric value specifying the size of the x-axis text. Defaults to 8. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_text_size}{A numeric value specifying the size of the y-axis text. Defaults to 12. See \code{\link[ggplot2]{element_text}}.}

\item{x_axis_title_size}{A numeric value specifying the size of the x-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{y_axis_title_size}{A numeric value specifying the size of the y-axis title. Defaults to 14. See \code{\link[ggplot2]{element_text}}.}

\item{plot_title_size}{A numeric value specifying the size of the plot title. Defaults to 20. See \code{\link[ggplot2]{element_text}}.}

\item{annotate_value_text_size}{A numeric value specifying the size of the annotation text. Defaults to 6. See \code{\link[ggplot2]{geom_text}}.}

\item{...}{Additional arguments that can be passed to \code{\link[nlraa]{predict_nlme}} or \code{\link[nlraa]{predict_nls}}}
}
\value{
When return = "plot", returns either a plot of the growth model's fixed-effects predicted values as a line and 95\% bootstrapped confidence intervals as a shaded region.
When return = "estimates", return a data frame with fixed-effects predicted values and 95\% bootstrapped confidence intervals for each observation.
}
\description{
The function leverages functions from the 'nlraa' package to create bootstrapped confidence intervals of the fixed-effects of a growth model fit from the function \code{\link{growth_curve_model_fit}} and summarized by \code{\link{summarize_growth_model}}.
}
\examples{
# Load example data (exponential data)
data(exp_mixed_data)
# Fit an mixed-effects growth model to the data
exp_mixed_model <- growth_curve_model_fit(data_frame = exp_mixed_data, function_type = "exponential")
# Summarize the data by creating a summary list object
exp_mixed_model_summary <- summarize_growth_model(data_frame = exp_mixed_data, growth_model_object = exp_mixed_model, model_type = "mixed", function_type = "exponential")
# Create a plot of fixed-effects bootstrapped 95\%CI with the doubling time estimates from the exp_mixed_model annotated on the plot
exp_ci_plot <- growth_model_boot_ci_curve(growth_model_object = exp_mixed_model,
growth_model_summary_list = exp_mixed_model_summary,
model_type = "mixed",
annotate_value = "double_time",
return = "plot")
print(exp_ci_plot)
}
\seealso{
\code{\link{growth_curve_model_fit}} \code{\link{summarize_growth_model}}
}
