% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_model_fixed_cor_table.R
\name{growth_model_fixed_cor_table}
\alias{growth_model_fixed_cor_table}
\title{Create a printable table of the growth model fixed-effects correlation matrix (mixed-effects models ONLY)}
\usage{
growth_model_fixed_cor_table(growth_model_summary_list, use_knit_print = FALSE)
}
\arguments{
\item{growth_model_summary_list}{A list object created by the \code{\link{summarize_growth_model}} function.}

\item{use_knit_print}{A logical value to specify whether the flextable should be printer using the \code{\link[knitr]{knit_print}} function instead of the flextable object being returned.
Defaults to FALSE.}
}
\value{
A flextable object of the 'model_fixed_corr' data frame
}
\description{
'growth_model_fixed_cor_table()' creates a flextable object that can be used for documentation or Rmarkdown reports from the list object created by \code{\link{summarize_growth_model}}.
The 'model_fixed_corr' data frame from the list object for mixed-effects models is used to generate the table.
}
\examples{
# Load example data (exponential data)
data(exp_mixed_data)
# Fit an mixed-effects growth model to the data
exp_mixed_model <- growth_curve_model_fit(data_frame = exp_mixed_data, function_type = "exponential")
# Summarize the data by creating a summary list object
exp_mixed_model_summary <- summarize_growth_model(data_frame = exp_mixed_data,
growth_model_object = exp_mixed_model,
model_type = "mixed",
function_type = "exponential")
# Create flextable object of the growth model fixed-effects correlation matrix ('model_fixed_corr' data frame from summary list object) for documentation
exp_model_cor_table <- growth_model_summary_table(growth_model_summary_list = exp_mixed_model_summary)
print(exp_model_cor_table)
}
\seealso{
\code{\link{growth_curve_model_fit}} \code{\link{summarize_growth_model}} \code{\link[flextable]{flextable}}
}
