% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz_mixed_model.R
\name{gompertz_mixed_model}
\alias{gompertz_mixed_model}
\title{Fit a Gompertz mixed-effects regression model}
\usage{
gompertz_mixed_model(data_frame, model_type = "mixed", fixed_rate = TRUE)
}
\arguments{
\item{data_frame}{A data frame object that at minimum contains three variables:
\itemize{
\item cluster - a character type variable used to specify how observations are nested or grouped by a particular cluster. Note if using a least-squares model, please fill in all values of cluster with a single dummy character string, do NOT leave blank.
\item time - a numeric type variable used for measuring time such as minutes, hours, or days
\item growth_metric - a numeric type variable used for measuring growth over time such as cell count or confluency
}}

\item{model_type}{A character string specifying the type of regression model to be used. If "mixed" a mixed-effects regression model will be used with fixed and random effects to account for clustering. If "least-squares". Defaults to "mixed".}

\item{fixed_rate}{A logical value specifying whether the rate constant of the function should be treated as a fixed effect (TRUE) or random effect (FALSE). Defaults to TRUE}
}
\value{
Returns a Gompertz model object of class 'nlme' when a mixed-effects model is specified or a model object of class 'nls' if a least-squares model is specified.
}
\description{
'gompertz_mixed_model()' is a function utilized with the \code{\link{growth_curve_model_fit}} function for fitting a Gompertz mixed-effects regression model to growth data utilizing the \code{\link[nlme]{nlme}} function.
Starting values are derived from an initial least-squares model using the \code{\link[minpack.lm]{nlsLM}} function.
}
\examples{
# Load example data (Gompertz data from GrowthCurveME package)
data(gomp_mixed_data)
# Fit a Gompertz mixed-effects growth model to the data using the main growth_curve_model_fit() function
gomp_mixed_model <- growth_curve_model_fit(data_frame = gomp_mixed_data, function_type = "gompertz")
# Fit a Gompertz mixed-effected model using the gompertz_mixed_model() function
gomp_mixed_model <- gompertz_mixed_model(data_frame = gomp_mixed_data)
}
\seealso{
\code{\link{growth_curve_model_fit}}
}
