% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curve_model_fit.R
\name{growth_curve_model_fit}
\alias{growth_curve_model_fit}
\title{Fit a growth function using mixed-effects regression modeling}
\usage{
growth_curve_model_fit(
  data_frame,
  function_type = "exponential",
  model_type = "mixed",
  fixed_rate = TRUE,
  num_chains = 1,
  time_unit = "hours",
  return_summary = TRUE,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_frame}{A data frame object that at minimum contains three
variables:
\itemize{
\item cluster - a character type variable used to specify how observations
are nested or grouped by a particular cluster. Note if using a
least-squares model, please fill in cluster values with a single repetitive
dummy variable (e.g., '1'), do not leave blank.
\item time - a numeric type variable used for measuring time such as
minutes, hours, or days
\item growth_metric - a numeric type variable used for measuring growth
over time such as cell count or confluency
}}

\item{function_type}{A character string specifying the function for
modeling the shape of the growth. Options include "exponential", "linear",
"logistic", or "gompertz".}

\item{model_type}{A character string specifying the type of regression
model to be used. If 'mixed', a mixed-effects regression model will be used
with fixed and random-effects to account for clustering. For
'least-squares', a least-squares regression model with only fixed-effects
is applied. Defaults to "mixed".}

\item{fixed_rate}{A logical value specifying whether the rate constant
of the function should be treated as a fixed effect (TRUE) or random
effect (FALSE). Defaults to TRUE}

\item{num_chains}{A numeric value specifying the number of chains to run
in parallel in the MCMC algorithm of saemix. Increasing the number of chains
may improve convergence but may also increase the computational time.
Defaults to 1.}

\item{time_unit}{A character string specifying the units in which time is
measured in. Defaults to "hours"}

\item{return_summary}{A logical value specifying whether to return the
growth_model_summary_list when TRUE (list object containing summarized data)
or the object model object when FALSE. Defaults to TRUE.}

\item{seed}{A numeric value specifying a seed number to reproduce the
random starting values sampled within the function. Defaults to NULL.}

\item{verbose}{A logical value specifying whether print statements will
print in the console. Defaults to TRUE.}
}
\value{
A list object with the following data frames within the list:
\itemize{
\item model_summary_wide - a data frame with 1 row containing
key model estimates, doubling-time, and model metrics depending
on the model_type and function_type specified
\item model_summary_long - a data frame that is a long dataset version of
'model_summary_wide' that can be used to generate a table of the model
results (see function \code{\link{growth_model_summary_table}})
\item model_residual_data - a data frame containing the original data
frame values as well as predicted values, residuals, and theoretical
quantiles of the residuals depending on the model_type selected
(see functions \code{\link{growth_model_residual_plots}} and
\code{\link{growth_vs_time_plot}})
\item simulated_data - a data frame with the 95\% prediction intervals
calculated using the median for the estimate and the 2.5th and 97.5th
percentiles of the simulated data at each time point (not to be
confused with the 95\% confidence intervals calculated from the model
estimates). See \code{\link{summarize_growth_model}}.
}
Note when return_summary is FALSE, will return a model object of class
'SaemixObject' when a mixed-effects model is specified or a model object of
class 'nls' if a least-squares model is specified.
}
\description{
This function fits a mixed-effects model to a data frame based
on a user-defined function to account for clustering.
}
\examples{
\donttest{
# Load example data (exponential data)
data(exp_mixed_data)
# Fit an mixed-effects growth model to the data and return summary
exp_mixed_model_summary <- growth_curve_model_fit(
data_frame = exp_mixed_data,
function_type = "exponential",
verbose = FALSE)
# Create flextable object from the summary list object for documentation
exp_model_table <- growth_model_summary_table(
growth_model_summary_list = exp_mixed_model_summary)
exp_model_table
# Create growth vs time plot of data with fitted values (plot_type = 2),
# adjust aesthetics and parameters as desired
exp_growth_plot <- growth_vs_time_plot(
growth_model_summary_list = exp_mixed_model_summary,
plot_type = 2)
print(exp_growth_plot)
# Check residuals and model assumptions
residual_diag_plot <- growth_model_residual_plots(
growth_model_summary_list = exp_mixed_model_summary)
print(residual_diag_plot)
}
}
